Module ejdb::bson

Modules

oid

ordered

spec
Constants derived from the BSON Specification Version 1.0.

Structs

Decoder

Encoder

Enums

Bson
Possible BSON value types.

DecoderError
Possible errors that can arise during decoding.

EncoderError
Possible errors that can arise during encoding.

ValueAccessError
Error to indicate that either a value was empty or it contained an unexpected type, for use with the direct getters.

Functions

decode_document
Attempt to decode a Document from a byte stream.

encode_document
Attempt to encode a Document into a byte stream.

from_bson
Decode a BSON Value into a T Deserializable.

to_bson
Encode a T Serializable into a BSON Value.

Type Definitions

Array

DecoderResult

Document

EncoderResult

ValueAccessResult