Struct ejdb::bson::ordered::OrderedDocument

pub struct OrderedDocument {
    // some fields omitted
}

A BSON document represented as an associative HashMap with insertion ordering.

Methods

impl OrderedDocument

fn new() -> OrderedDocument

Creates a new empty OrderedDocument.

fn iter(&'a self) -> OrderedDocumentIterator<'a>

Gets an iterator over the entries of the map.

fn clear(&mut self)

Clears the document, removing all values.

fn get(&self, key: &str) -> Option<&Bson>

Returns a reference to the Bson corresponding to the key.

fn get_mut(&mut self, key: &str) -> Option<&mut Bson>

Gets a mutable reference to the Bson corresponding to the key

fn get_f64(&self, key: &str) -> Result<f64, ValueAccessError>

Get a floating point value for this key if it exists and has the correct type.

fn get_str(&self, key: &str) -> Result<&str, ValueAccessError>

Get a string slice this key if it exists and has the correct type.

fn get_array(&self, key: &str) -> Result<&Vec<Bson>, ValueAccessError>

Get a reference to an array for this key if it exists and has the correct type.

fn get_document(&self, key: &str) -> Result<&OrderedDocument, ValueAccessError>

Get a reference to a document for this key if it exists and has the correct type.

fn get_bool(&self, key: &str) -> Result<bool, ValueAccessError>

Get a bool value for this key if it exists and has the correct type.

fn is_null(&self, key: &str) -> bool

Returns wether this key has a null value

fn get_i32(&self, key: &str) -> Result<i32, ValueAccessError>

Get an i32 value for this key if it exists and has the correct type.

fn get_i64(&self, key: &str) -> Result<i64, ValueAccessError>

Get an i64 value for this key if it exists and has the correct type.

fn get_time_stamp(&self, key: &str) -> Result<i64, ValueAccessError>

Get a time stamp value for this key if it exists and has the correct type.

fn get_binary_generic(&self, key: &str) -> Result<&Vec<u8>, ValueAccessError>

Get a generic binary value for this key if it exists and has the correct type.

fn get_object_id(&self, key: &str) -> Result<&ObjectId, ValueAccessError>

Get an object id value for this key if it exists and has the correct type.

fn get_utc_datetime(&self, key: &str) -> Result<&DateTime<UTC>, ValueAccessError>

Get a UTC datetime value for this key if it exists and has the correct type.

fn contains_key(&self, key: &str) -> bool

Returns true if the map contains a value for the specified key.

fn keys(&'a self) -> Keys<'a>

Gets a collection of all keys in the document.

fn values(&'a self) -> Values<'a>

Gets a collection of all values in the document.

fn len(&self) -> usize

Returns the number of elements in the document.

fn is_empty(&self) -> bool

Returns true if the document contains no elements

fn insert<KT, BT>(&mut self, key: KT, val: BT) -> Option<Bson> where BT: Into<Bson>, KT: Into<String>

Sets the value of the entry with the OccupiedEntry's key, and returns the entry's old value. Accepts any type that can be converted into Bson.

fn insert_bson(&mut self, key: String, val: Bson) -> Option<Bson>

Sets the value of the entry with the OccupiedEntry's key, and returns the entry's old value.

fn remove(&mut self, key: &str) -> Option<Bson>

Takes the value of the entry out of the document, and returns it.