Struct ejdb::meta::CollectionMetadata

pub struct CollectionMetadata<'a>(_);

Represents metainformation about a collection in an EJDB database.

This structure only provides a view into the original BSON document, therefore it has a lifetime dependency on the database metadata structure. It is also not possible to unwrap the inner document, though it still has a Deref implementation with bson::Document as a target.

Methods

impl<'a> CollectionMetadata<'a>

fn name(&self) -> &str

Returns the name of this collection.

fn file(&self) -> &str

Returns the file path of this collection.

fn records(&self) -> u64

Returns the number of records in this collection.

fn buckets(&self) -> u64

Returns the number of buckets in this collection.

fn cached_records(&self) -> u64

Returns the number of cached records for this collection.

fn large(&self) -> bool

Returns true if the collection can hold more than 2GB of data, false otherwise.

fn compressed(&self) -> bool

Returns true if DEFLATE compression is applied to this collection's records, false otherwise.

fn indices(&self) -> CollectionIndices

Returns an iterator of metadata of indices in this collection.