Struct ejdb::meta::IndexMetadata

pub struct IndexMetadata<'a>(_);

Represents metainformation about an index of a collection in an EJDB database.

Like CollectionMetadata, this structure only provides a view into the full database metadata object, so it is not possible to obtain a bson::Document directly. Deref<Target=bson::Document> implementation is available.

Methods

impl<'a> IndexMetadata<'a>

pub fn field(&self) -> &str

Returns the name of the field on which this index is defined.

pub fn name(&self) -> &str

Returns the name of this index itself (usually it is automatically generated).

pub fn index_type(&self) -> IndexType

Returns the type of this index.

pub fn records(&self) -> Option<u64>

Returns the number of records using this index, if available.

pub fn file(&self) -> Option<&str>

Returns the path to the file of this index, if available.