Struct ejdb::query::FieldConstraint

pub struct FieldConstraint(_, _);

A transient builder for adding field-based query constraints.

Instances of this structure are returned by Query::field() and FieldConstrait::field() methods.

Methods

impl FieldConstraint

fn field<S: Into<String>>(self, name: S) -> FieldConstraint

Returns a constraint builder for a deeper field.

fn eq<V: Into<Bson>>(self, value: V) -> Query

Adds an equality constraint for this field and value.

fn begin<S: Into<String>>(self, value: S) -> Query

Adds a $begin constraint for this field.

The value of this field must start with value. The field type should be string.

fn between<N1: BsonNumber, N2: BsonNumber>(self, left: N1, right: N2) -> Query

Adds a $between constraint for this field.

The value of this field must be greater than or equal to left and less than or equal to right. The field type should be numeric.

fn gt<N: BsonNumber>(self, value: N) -> Query

Adds a $gt constraint for this field.

The value of this field must be strictly greater than value. The field type should be numeric.

fn gte<N: BsonNumber>(self, value: N) -> Query

Adds a $gte constraint for this field.

The value of this field must be greater than or equal to value. The field type should be numeric.

fn lt<N: BsonNumber>(self, value: N) -> Query

Adds an $lt constraint for this field.

The value of this field must be strictly less than value. The field type should be numeric.

fn lte<N: BsonNumber>(self, value: N) -> Query

Adds an $lte constraint for this field.

The value of this field must be less than or equal to value. The field type should be numeric.

fn exists(self, exists: bool) -> Query

Adds an $exists constraint for this field.

The field must exists if exists is true, the opposite otherwise.

fn elem_match<Q: Into<Document>>(self, query: Q) -> Query

Adds an $elemMatch constraint for this field.

Any element of the array contained in this field must match query. The query argument is a regular EJDB query; you can pass another Query object to it.

fn contained_in<I>(self, values: I) -> Query where I: IntoIterator, I::Item: Into<Bson>

Adds an $in constraint for this field.

The value of this field must be equal to one of the values yielded by values.

fn not_contained_in<I>(self, values: I) -> Query where I: IntoIterator, I::Item: Into<Bson>

Adds an $nin constraint for this field.

The value of this field must not be equal to all of the values yielded by values.

fn case_insensitive(self) -> FieldConstraint

Adds a $icase constraint for this field.

Makes all further constraints for this field case insensitive with regard to string values.

fn not(self) -> FieldConstraint

Adds a $not constraint for this field.

Inverts the further query, making only those elements which do NOT satisfy the following constraints to match.

fn str_and<I>(self, values: I) -> Query where I: IntoIterator, I::Item: Into<String>

Adds an $strand constraint for this field.

If this field holds an array of strings, $strand returns those records whose array contains all elements from values.

If this field holds a string, it is first split into an array by space ' ' or comma ',' characters, and the resulting array is queried like in 1.

fn str_or<I>(self, values: I) -> Query where I: IntoIterator, I::Item: Into<String>

Adds an $stror constraint for this field.

If this field holds an array of strings, $stror returns those records whose array contains at least one element from values.

If this field holds a string, it is first split into an array by space ' ' or comma ',' characters, and the resulting array is queried like in 1.