Struct ejdb::query::QH

pub struct QH;

An entry point for constructing query hints.

This is a convenience API. This structure provides the same methods as QueryHints structure and inside them a fresh QueryHints instance is created and the corresponding method is called on it. This is the main approach for constructing query hints.

Example

use ejdb::query::{QueryHints, QH};

assert_eq!(
    QueryHints::new().max(128).field("name").include().order_by("date").desc(),
    QH.max(128).field("name").include().order_by("date").desc()
)

Methods

impl QH

fn empty(self) -> QueryHints

fn max(self, n: i64) -> QueryHints

fn skip(self, n: i64) -> QueryHints

fn order_by<S: Into<String>>(self, field: S) -> QueryHintsOrderBy

fn field<S: Into<String>>(self, field: S) -> QueryHintsField