Struct ejdb::query::QueryHintsOrderBy

pub struct QueryHintsOrderBy(_, _);

A builder for ordering hint by a specific field.

Methods

impl QueryHintsOrderBy

fn desc(self) -> QueryHints

Sets that query results must be sorted by this field in descending order.

fn asc(self) -> QueryHints

Sets that query results must be sorted by this field in ascending order.