Struct ejdb::query::Q

pub struct Q;

An entry point for constructing queries.

This is a convenience API. This structure provides the same methods as Query structure and inside them a fresh Query instance is created and the corresponding method is called on it. This is the main approach for constructing queries.

Example

use ejdb::query::{Query, Q};

assert_eq!(
    Query::new().field("name").eq("Foo").inc("rating", 1).set("favorite", true),
    Q.field("name").eq("Foo").inc("rating", 1).set("favorite", true)
)

Methods

impl Q

pub fn empty(self) -> Query

pub fn and<I>(self, queries: I) -> Query where
I: IntoIterator,
I::Item: Into<Document>,

pub fn or<I>(self, queries: I) -> Query where
I: IntoIterator,
I::Item: Into<Document>,

pub fn id<V: Into<Bson>>(self, value: V) -> Query

pub fn field<S: Into<String>>(self, name: S) -> FieldConstraint

pub fn join<S: Into<String>, C: Into<String>>(self, key: S, coll: C) -> Query

pub fn add_to_set<S: Into<String>, V: Into<Bson>>(
self,
key: S,
value: V
) -> Query

pub fn add_to_set_all<S, I>(self, key: S, values: I) -> Query where
S: Into<String>,
I: IntoIterator,
I::Item: Into<Bson>,

pub fn unset<S: Into<String>>(self, key: S) -> Query

pub fn inc<S: Into<String>, D: BsonNumber>(self, key: S, delta: D) -> Query

pub fn drop_all(self) -> Query

pub fn upsert_many<D: Into<Document>>(self, document: D) -> Query

pub fn upsert<S: Into<String>, V: Into<Bson>>(self, key: S, value: V) -> Query

pub fn set<S: Into<String>, V: Into<Bson>>(self, key: S, value: V) -> Query

pub fn set_many<D: Into<Document>>(self, document: D) -> Query

pub fn pull<S: Into<String>, V: Into<Bson>>(self, key: S, value: V) -> Query

pub fn pull_all<S, I>(self, key: S, values: I) -> Query where
S: Into<String>,
I: IntoIterator,
I::Item: Into<Bson>,

pub fn push<S: Into<String>, V: Into<Bson>>(self, key: S, value: V) -> Query

pub fn push_all<S, I>(self, key: S, values: I) -> Query where
S: Into<String>,
I: IntoIterator,
I::Item: Into<Bson>,

pub fn rename<S1: Into<String>, S2: Into<String>>(
self,
key: S1,
new_key: S2
) -> Query

pub fn slice<S: Into<String>>(self, key: S, limit: i64) -> Query

pub fn slice_with_offset<S: Into<String>>(
self,
key: S,
offset: i64,
limit: i64
) -> Query