Struct ejdb::query::QueryHints

pub struct QueryHints { /* fields omitted */ }

A container of EJDB query options.

This structure is a wrapper around a BSON document with various options affecting query execution in EJDB. It implements Deref<Target=bson::Document> and DerefMut, therefore it is possible to work with it as a BSON document directly. It also has into_bson()/as_bson() methods and Into<bson::Document>/From<bson::Document> implementations. If an invalid document is constructed and passed as a hints map when executing a query, an error will be returned.

Query hints are a part of any query operation and are passed with the actual query to Collection::query() method; they can be empty if the default behavior is sufficient.

Methods

impl QueryHints

pub fn new() -> QueryHints

Creates a new, empty query hints set.

pub fn max(self, n: i64) -> QueryHints

Sets the maximum number of entries which should be returned by the query.

Corresponds to $max hint in EJDB query hints syntax.

pub fn skip(self, n: i64) -> QueryHints

Sets the number of entries which should be skipped first when query results are inspected.

Corresponds to $skip hint in EJDB query hints syntax.

pub fn order_by<S: Into<String>>(self, field: S) -> QueryHintsOrderBy

Returns a builder for ordering hint for the provided field.

Corresponds to $orderBy hint in EJDB query hints syntax.

pub fn field<S: Into<String>>(self, field: S) -> QueryHintsField

Returns a builder for setting inclusion/exclusion flag of the provided field.

Corresponds to $fields hint in EJDB query syntax.

pub fn into_bson(self) -> Document

Converts these hints to a BSON document.

pub fn as_bson(&self) -> &Document

Returns a reference to these hints as a BSON document.

pub fn as_bson_mut(&mut self) -> &mut Document

Returns a mutable reference to these hints as a BSON document.

Be careful when modifying the document directly because it may lead to invalid hints.