Struct ejdb::query::QueryHintsOrderBy

pub struct QueryHintsOrderBy(_, _);

A builder for the hint to order by a specific field.

Methods

impl QueryHintsOrderBy

pub fn desc(self) -> QueryHints

Sets that query results must be sorted by this field in descending order.

pub fn asc(self) -> QueryHints

Sets that query results must be sorted by this field in ascending order.